refactor of badly placed selects, moved to pojos
